pyproject.toml
[project]
name = "tesseract-hot-folder-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "Scott Hendrickson",email = "scott@drskippy.net"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
